requirements.txt
Django==5.1.6
Pillow==11.1.0
PyMySQL==1.1.1
cryptography==44.0.1
djangorestframework==3.15.2
sqlacodegen==3.0.0
django-rest-authoken=2.1.4